requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-jaeger
opentelemetry-exporter-prometheus
